+++ src/main/java/commandHelp/Command.java
@@ -2,7 +2,7 @@ package commandHelp;
 
 // With some patience you should replace the hardcoded command names with these. 
 public enum Command {
-	featureColorForRegex("featureColorForRegex"), 
+	featureColor("featureColor"), 
 	featureDisplayMode("featureDisplayMode"), 
 	print("print"),
 	q("q"),