+++ src/main/java/commandHelp/CommandList.java
@@ -622,22 +622,26 @@ public class CommandList {
 		cmdList.add(cmd);
 		
 		cmd= new CommandHelp();
-		cmd.setName("featureColorForRegex"); cmd.setArgs("[-r/-R regex color] [-v] [track_regex = .*]..."); cmd.inSection= Section.DISPLAY; 
-		cmd.setBriefDescription("Set colour for features captured by regex. ");
+		cmd.setName("featureColor"); cmd.setArgs("[-r/-R expression color] [-v] [track_regex = .*]..."); cmd.inSection= Section.DISPLAY; 
+		cmd.setBriefDescription("Set colour for features captured by expression. ");
 		cmd.setAdditionalDescription(""
-				+ "This command affects interval feature tracks (bed, gff, vcf, etc) and overrides the default color "
+				+ "This command affects interval feature tracks (bed, gff, vcf, etc) "
+				+ "and overrides the default color "
 				+ "for the lines captured by "
-				+ "a regex. It is useful to highlight features containg a string of interest, "
-				+ "such as 'CDS' in gff files.\n"
+				+ "the expression. Expression is a regex or an awk script (autodetermined). "
+				+ "It is useful to highlight features containg a "
+				+ "string of interest, "
+				+ "such as 'CDS' in gff files, or features where a numeric field satisfy a filter.\n"
 				+ "\n"
 				+ "Options:\n"
 				+ "\n"
-				+ ":code:`-r <regex> <color>` Features matching :code:`regex` will have color :code:`color`. "
-				+ "The regex is applied to the raw lines as read from file. "
-				+ "This option takes exactly two arguments and can be given zero or more times. "
-				+ "If this option is not present colors are reset to default.\n"
+				+ ":code:`-r <expression> <color>` Features matching :code:`expression` will "
+				+ "have color :code:`color`. "
+				+ "The expression is interpreted as regex or as an awk script and it "
+				+ "is applied to the raw lines as read from file. "
+				+ "This option takes exactly two arguments and can be given zero or more times.\n"
 				+ "\n"
-				+ ":code:`-R <regex> <color>` Same as :code:`-r` but sets color for features NOT "
+				+ ":code:`-R <expression> <color>` Same as :code:`-r` but sets color for features NOT "
 				+ "matched by regex.\n"
 				+ "\n"
 				+ ":code:`-v` Invert selection: apply changes to the tracks not selected by list of track_regex\n"
@@ -646,11 +650,15 @@ public class CommandList {
 				+ "\n"
 				+ "Example::\n"
 				+ "\n"
-				+ "    featureColorForRegex -r CDS plum2 -r exon grey\n"
-				+ "    featureColorForRegex bed~~~~~~~~~-> Reset to default the track matching 'bed'\n"
-				+ "	   featureColorForRegex -R CDS grey -> Grey all features except those matching CDS\n"
+				+ "    featureColor -r CDS plum2 -r exon grey\n"
+				+ "    featureColor bed~~~~~~~~~-> Reset to default the track matching 'bed'\n"
+				+ "	   featureColor -R CDS grey -> Grey all features except those matching CDS\n"
+				+ "    \n"
+				+ "    Color blue where 9th field is > 3; color red where 9th is > 6\n"
+				+ "    featureColor -r '$9 > 3' blue -r '$9 > 6' red\n"
 				+ "\n"
-				+ "Colors can be specified by name, name prefix, or integer in range 0-255. Available colours:\n"
+				+ "Colors can be specified by name, name prefix, or integer in range 0-255. "
+				+ "Available colours:\n"
 				+ "\n"
 				+ Xterm256.colorShowForTerminal().replaceAll(" ", "~") 
 				+ "\n");
@@ -1176,11 +1184,13 @@ public class CommandList {
 		cmd.setName("orderTracks"); cmd.setArgs("[track_regex]..."); cmd.inSection= Section.GENERAL; 
 		cmd.setBriefDescription("Reorder tracks according to the list of regexes or sort by name.");
 		cmd.setAdditionalDescription("Not all the tracks need to be listed, the missing ones "
-				+ "follow the listed ones in unchanged order. Without arguments sort track by tag name.\n"
+				+ "follow the listed ones in unchanged order. Without arguments sort "
+				+ "track by tag name.\n"
 				+ "For example, given the track list: `[hela.bam#1, hela.bed#2, hek.bam#3, hek.bed#4]`::\n"
 				+ "\n"
 				+ "    orderTracks #2 #1~~~->~[hela.bed#2, hela.bam#1, hek.bam#3, hek.bed#4]\n"
 				+ "    orderTracks bam bed~->~[hela.bam#1, hek.bam#3, hela.bed#2, hek.bed#4]\n"
+				+ "    orderTracks . bam ~->~'bam' tracks go last\n"
 				+ "    orderTracks~~~~~~~~~->~name sort~[hela.bam#1, hela.bed#2, hek.bam#3, hek.bed#4]\n"
 				+ "");
 		cmdList.add(cmd);
@@ -1219,8 +1229,8 @@ public class CommandList {
 		cmd= new CommandHelp();
 		cmd.setName("samtools"); cmd.setArgs("[-f INT=0] [-F INT=4] [-q INT=0] [-v] [track_re = .*] ..."); cmd.inSection= Section.ALIGNMENTS; 
 		cmd.setBriefDescription("Apply samtools filters to alignment tracks captured by the list of track regexes.");
-		cmd.setAdditionalDescription("As *samtools view*, this command filters alignment records on the basis "
-				+ "of the given flags:\n"
+		cmd.setAdditionalDescription("Of interest to stranded RNA-Seq and BS-Seq, the bit flag 4096 is internally specified "
+				+ "to selects reads mapping to TOP STRAND."
 				+ "\n"
 				+ "* :code:`-F` Filter out flags with these bits set. NB: 4 is always set.\n"
 				+ "\n"
@@ -1234,6 +1244,7 @@ public class CommandList {
 				+ "\n"
 				+ "    samtools -q 10~~~~~~~~~~~-> Set mapq for all tracks. -f and -F reset to default\n"
 				+ "    samtools -F 1024 foo bar -> Set -F for all track containing re foo or bar\n"
+				+ "    samtools -f 4096 ~~~~~~~~-> Select TOP STRAND reads\n"
 				+ "    samtools~~~~~~~~~~~~~~~~~-> Reset all to default.\n"
 				+ "");
 		cmdList.add(cmd);
@@ -1384,7 +1395,7 @@ public class CommandList {
 		paramList.add("gap");
 		paramList.add("trackHeight");
 		paramList.add("colorTrack");
-		paramList.add("featureColorForRegex");
+		paramList.add("featureColor");
 		paramList.add(Command.featureDisplayMode.getCmdDescr());
 		paramList.add("readsAsPairs");
 		paramList.add("hideTitle");