+++ src/main/java/tracks/TrackReads.java
@@ -58,7 +58,7 @@ public class TrackReads extends Track{
 		this.setTrackFormat(TrackFormat.BAM);
 		
 		if(!Utils.bamHasIndex(bam)){
-			File temp= Utils.createTempFile(".asciigenome.", ".bam");
+			File temp= Utils.createTempFile(".asciigenome.", ".bam", true);
 			Utils.sortAndIndexSamOrBam(bam, temp.getAbsolutePath(), true);
 			this.setWorkFilename(temp.getAbsolutePath());
 		} else {